+++ src/matching_engine/event/fill_event.cpp
@@ -1,6 +1,6 @@
 #include "matching_engine/event/fill_event.h"
 
-FillEvent::FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, double qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {
+FillEvent::FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, int64_t qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {
     if (price <= 0) {
         throw std::invalid_argument("Price must be greater than 0");
     }
@@ -13,4 +13,4 @@ EventType FillEvent::getType() const { return EventType::Fill; }
 
 uint64_t FillEvent::getMatchId() const { return _matchId; }
 int64_t FillEvent::getPrice() const { return _price; }
-double FillEvent::getQty() const { return _qty; }
+int64_t FillEvent::getQty() const { return _qty; }
