+++ src/matching_engine/order/order.cpp
@@ -1,6 +1,6 @@
 #include "matching_engine/order/order.h"
 
-Order::Order(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, int64_t price, double qtyRemaining) : _clientId(clientId), _orderId(orderId), _symbol(symbol), _side(side), _price(price), _qtyRemaining(qtyRemaining) {
+Order::Order(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, int64_t price, int64_t qtyRemaining) : _clientId(clientId), _orderId(orderId), _symbol(symbol), _side(side), _price(price), _qtyRemaining(qtyRemaining) {
     if (price <= 0) {
         throw std::invalid_argument("Price must be greater than 0");
     }
@@ -17,10 +17,10 @@ uint64_t Order::getOrderId() const { return _orderId; }
 const std::string& Order::getSymbol() const & { return _symbol; }
 Side Order::getSide() const { return _side; }
 int64_t Order::getPrice() const { return _price; }
-double Order::getQtyRemaining() const { return _qtyRemaining; }
+int64_t Order::getQtyRemaining() const { return _qtyRemaining; }
 bool Order::isFilled() const { return _isFilled; }
 
-double Order::qtyDecrease(double execQty) {
+int64_t Order::qtyDecrease(int64_t execQty) {
     if (execQty <= 0) {
         throw std::invalid_argument("Executed quantity must be greater than 0");
     }
